+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.70.
+## 0.9.0
+
+- Rename `spec::{Schema => ObjectSchema}` struct.
+- Add `spec::BooleanSchema` struct.
+- Add `spec::Schema` enum.
+- The `spec::ObjectSchema::addition_properties` field is now of type `Option<Schema>`.
+- The `spec::Parameter::schema` field is now of type `ObjectOrReference<ObjectSchema>`.
+- Add `Operation::extensions` field.
+- Minimum supported Rust version (MSRV) is now 1.75.
 
 ## 0.8.1